ChatNow/requirements.txt
streamlit>=1.35
python-dotenv
xai-sdk>=1.5.0
pdfplumber
python-docx
langchain-text-splitters
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers